refactor  ESbuild and other minor things
- Add format to esbuild build params (not sure if CommonJS is the default)
- Add a dev script
- Remove tsc package
- Add Events enum export in types folder
- Add dist folder to exclude option in tsconfig, allows overwrite of types
