+++ src/types/index.ts
@@ -9,3 +9,17 @@ export interface CacheItem {
 export type TStrings = string[];
 export type TKey = string | TStrings;
 export type TMapCache = Map<string, CacheItem>;
+
+export enum Events {
+  ON_SET = 'onSet',
+  ON_HIT = 'onHit',
+  ON_MISS = 'onMiss',
+  ON_TTL_EXPIRED = 'onTtlExpired',
+  ON_INVALIDATED = 'onInvalidated',
+  ON_DELETED = 'onDeleted',
+  ON_CLEAR = 'onClear',
+  ON_SNAPSHOT_RESTORED = 'onSnapshotRestored',
+  ON_SNAPSHOT_RESTORE_FAILED = 'onSnapshotRestoreFailed'
+}
+
+export type { CacheStrategy } from './CacheStrategy';
