+++ tsconfig.json
@@ -10,5 +10,5 @@
     "moduleResolution": "node",
     "outDir": "dist"
   },
-  "exclude": ["node_modules", "**/*.test.js"]
+  "exclude": ["node_modules", "**/*.test.js", "dist"]
 }
